+++ HEJZ_front/src/App.tsx
@@ -16,7 +16,7 @@ import StatsScreen from './screens/StatesScreen';
 import SelectScreen from './screens/SelectScreen';
 import SignUpScreen from './screens/SignUpScreen';
 import SongPlayScreen from './screens/Community/SongPlayScreen';
-import TestVideo from './screens/TestVideo';
+
 
 
 enableScreens(); 
@@ -54,7 +54,7 @@ const App = () => {
                   <Stack.Screen name="Dance" component={DanceNavigator} />
                   <Stack.Screen name="Community" component={CommunityNavigator} />
                   <Stack.Screen name="SongPlay" component={SongPlayScreen} />
-                  <Stack.Screen name="TestVideo" component={TestVideo} />
+
 
                 </Stack.Navigator>
